Fix the syle of the "small-para" class
The <p> with the "small-para" class should had a
smaller font size. The rule to achieve this was
written before another rule that selected the
"para" class and applied a bigger size to it.

Because the <p> element in question also had a
"para" class, the problem was caused by the order
of the rules.

To solve the problem I deleted the last rule, and
moved the bigger font size property value under a
much previous rule that grouped both the "para"
and "small-para" classes.